fix(ui): Wrap renderer in liveness scope instead of function element
